fixes issue 258
git-svn-id: https://get-simple-cms.googlecode.com/svn/trunk@553 c6e34258-837a-11de-9f42-234e1f44043b

+++ admin/setup.php
@@ -111,16 +111,18 @@ if(isset($_POST['submitted'])) {
 		}
 
 		# create root .htaccess file
-		$init = GSROOTPATH.'.htaccess';
-		$temp_data = file_get_contents(GSROOTPATH .'temp.htaccess');
-		$temp_data = str_replace('**REPLACE**',tsl($path_parts), $temp_data);
-		$fp = fopen($init, 'w');
-		fwrite($fp, $temp_data);
-		fclose($fp);
-		if (!file_exists($init)) {
-			$kill .= sprintf(i18n_r('ROOT_HTACCESS_ERROR'), 'temp.htaccess', '**REPLACE**', tsl($path_parts)) . '<br />';
-		} else {
-			unlink(GSROOTPATH .'temp.htaccess');
+		if(in_arrayi('mod_rewrite',apache_get_modules())) {
+			$init = GSROOTPATH.'.htaccess';
+			$temp_data = file_get_contents(GSROOTPATH .'temp.htaccess');
+			$temp_data = str_replace('**REPLACE**',tsl($path_parts), $temp_data);
+			$fp = fopen($init, 'w');
+			fwrite($fp, $temp_data);
+			fclose($fp);
+			if (!file_exists($init)) {
+				$kill .= sprintf(i18n_r('ROOT_HTACCESS_ERROR'), 'temp.htaccess', '**REPLACE**', tsl($path_parts)) . '<br />';
+			} else {
+				unlink(GSROOTPATH .'temp.htaccess');
+			}
 		}
 		
 		# create gsconfig.php if it doesn't exist yet.
